pyproject.toml
[project]
name = "lab-template-backend"
version = "0.1.0"
description = "Backend for lab template system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
]
